+++ packages/components/quotation/index.js
@@ -1,2 +1,3 @@
 export * from './modals/index.js';
 export * from './views/index.js';
+export { mountQuotationFlow } from './logic/createQuotationFlowComponent.js';